live-api/app/server/requirements.txt
google-generativeai
google-cloud-aiplatform
google-adk
websockets
python-dotenv
llama_index
langgraph
langchain-tavily
